+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Experimental. Do not use in production.**
 
+* Change plugin name to webhookreceiver
 * Add webhook_receiver service
 * Ensure Bulk Enrollment is activated
 * Add basic testing via tox